feat(entrypoint.sh): enhance schedule calculation using cron schedule from environment variable
Refactor the calculation of the next scheduled time to use the cron
schedule from an environment variable. This change allows for more
flexible scheduling by parsing the cron format and calculating the
next execution time using busybox date. This also ensures that if the
calculated time is in the past, a day is added to the schedule.